f4f4 mxfp4-V: v_descale is a uint8 E8M0 image (drop 512B fp32 header)
_pack_v_mxfp4_colmajor now returns v_descale as the uint8 E8M0 block-scale
image [b, h_kv, nT*512] with no fp32 per-channel header. C++ fmha_v3_fwd_f4f4
asserts a byte dtype (was float32); the kernel self-computes the per-(b,h)
slice (stride kv_seq_len*4). bench_sage's f4f4 runner keeps v_descale uint8
(a .to(float32) would value-cast and corrupt the E8M0 bytes). Rebuilt
fwd_hd128_f4f4.co; cosine vs bf16 unchanged (~0.962).

+++ op_tests/op_benchmarks/triton/bench_sage.py
@@ -648,7 +648,9 @@ def make_f4f4_runner(
     k_quant = k_quant.contiguous()
     q_descale = q_descale.contiguous()
     k_descale = k_descale.contiguous()
-    v_descale = v_descale.to(torch.float32).contiguous()
+    # f4f4 v_descale is the uint8 E8M0 image (no fp32 header) -- keep it uint8; a .to(float32)
+    # here would VALUE-cast the E8M0 bytes and corrupt the block scales.
+    v_descale = v_descale.contiguous()
 
     softmax_scale = ASM_SPARSE_HEAD_DIM ** -0.5
